+++ default-config/main.go
@@ -64,7 +64,9 @@ func run(ctx context.Context, cfg jobConfig) error {
 	doguConfigRepo := repository.NewDoguConfigRepository(k8sConfigMapClient)
 	sensitiveDoguConfigRepo := repository.NewSensitiveDoguConfigRepository(k8sSecretClient)
 
-	ca := config.NewDefaultConfigApplier(globalConfigRepo, doguConfigRepo, sensitiveDoguConfigRepo, k8sSecretClient)
+	initialDomain := os.Getenv("INITIAL_DOMAIN")
+
+	ca := config.NewDefaultConfigApplier(globalConfigRepo, doguConfigRepo, sensitiveDoguConfigRepo, k8sSecretClient, initialDomain)
 	fa := fqdn.NewApplier(globalConfigRepo, k8sServicesClient)
 
 	if err = applyDefaults(ctx, cfg, ca, fa); err != nil {