+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#56] Option `defaultConfig.env.initialDomain` in Helm values to set a custom domain
+  - This change is necessary because once we have the LOP-IDP component,
+    we need to have the correct domain set when it is installed because we do not support domain changes in LDAP.
+### Changed
+- Default for the domain in the global config is now `ces.localdomain` as `*.local` should be used for mDNS only.
 
 ## [v4.0.0] - 2026-04-08
 ### Changed